Created an exception specifcially for decoding, in a move toward eliminating all of the non-specific 'throws exception' invocations that are all over the place.  This allowed some of these events (decoding in particular) to be more easily differentiable by the client.  (The client is still free to "catch(Exception)" if they so desire so this shouldn't cause any problems for existing code).
Modified some of the sending functions so that they catch ExecutionExceptions and re-throw them as IOExceptions if appropriate.  This allows client software to differentiate out the type of exception that would warrant a reconnection attempt / retry.

Most of this is to counter the fact that AsynchronousSocketChanenldoes not do very good connection tracking, and all the read() etc. calls return Future<T> placing any exceptions a layer a layer deeper.  It is inconvenient for the client code to have to deal with this.